First half of "make it work like chroot" work done
The idea is to make it possible to very simply leave
daemons etc running and return to the shell. This requires
the nschrooter to provide an init-substitute for them, or
else the quit of the program would also kill all the daemons.

(The second half is to use the .pid1-file to be able to transparently
re-enter the same namespace on subsequent uses of nschrooter.)