Add a timeout control for exiting an empty namespace
It might be that you want to keep an ns around even without
any non-init processes in it, eg. for mounts, or during
scripts with a bunch of "chroot" (replaced by nschrooter)
commands in it, the default 5 secs is for scripts,
the -1 = forever is in case you want to keep mounts.