nschrooter: Implement seccomp-based syscall filtering for chown,set*id
Makes for less messy running of stuff in user namespaces
where you never actually want to setuid or chown stuff.

Nssu gains a check for actually changing uid on setuid because
this filtering cheated it too to think that we actually are
privileged (this is actually good, cheats more things :P).